+++ libs/components/src/lib/text-field/ui.test.ts
@@ -31,7 +31,8 @@ test('should show the component', async ({page}: { page: Page }) => {
 	});
 
 	const testWrapper = await page.$('#wrapper');
-
+	await page.locator('vwc-text-field').nth(0).focus();
+	await page.keyboard.press('Tab');
 	await page.waitForLoadState('networkidle');
 
 	expect(await testWrapper?.screenshot())