+++ libs/components/src/visual-tests/visual-tests-utils.ts
@@ -95,15 +95,38 @@ export async function loadTemplate({
 	page,
 	template,
 }: { page: Page, template: string }) {
-	const browserType = page.context()?.browser()?.browserType().name();
-
 	const style = '';
 
 	await page.$('html').then(html => html?.evaluate((html) => {
 		html.classList.add('vvd-root');
 	}, template));
 
-	await page.$('body').then(body => body?.evaluate((body, {template, style}) => {
-		body.innerHTML = `${style}<div id="wrapper">${template}</div>`;
-	}, {template, style}));
+	await page.$('body').then((body) =>
+		body?.evaluate(
+			(body, { template, style }) => {
+				body.innerHTML = `${style}<div id="wrapper">${template}</div>`;
+			},
+			{ template, style }
+		)
+	);
+}
+
+// See https://github.com/microsoft/playwright/issues/6347
+export async function useFakeTime(page: Page, fakeNow: number) {
+	await page.addInitScript(`{
+		// Extend Date constructor to default to fakeNow
+		Date = class extends Date {
+			constructor(...args) {
+				if (args.length === 0) {
+					super(${fakeNow});
+				} else {
+					super(...args);
+				}
+			}
+		}
+		// Override Date.now() to start from fakeNow
+		const __DateNowOffset = ${fakeNow} - Date.now();
+		const __DateNow = Date.now;
+		Date.now = () => __DateNow() + __DateNowOffset;
+	}`);
 }